Removed seperate controllers
