+++ core/templates/dev/head/pages/suggestion_editor/ShowSuggestionModalForLearnerViewService.ts
@@ -17,6 +17,7 @@
  */
 
 require('domain/utilities/UrlInterpolationService.ts');
+require('pages/suggestion_editor/SuggestionModalService.ts');
 
 oppia.factory('ShowSuggestionModalForLearnerViewService', [
   '$rootScope', '$uibModal', 'UrlInterpolationService',
@@ -42,7 +43,19 @@ oppia.factory('ShowSuggestionModalForLearnerViewService', [
             return description;
           }
         },
-        controller: 'ShowSuggestionModalForLearnerView'
+        controller: [
+          '$scope', '$uibModalInstance', 'SuggestionModalService',
+          'description', 'newContent', 'oldContent',
+          function($scope, $uibModalInstance, SuggestionModalService,
+              description, newContent, oldContent) {
+            $scope.newContent = newContent;
+            $scope.oldContent = oldContent;
+            $scope.description = description;
+            $scope.cancel = function() {
+              SuggestionModalService.cancelSuggestion($uibModalInstance);
+            };
+          }
+        ]
       });
     };
